Fix sale date getter
In order to be coherent with the searcher method the sale_date of the invoice
must be the max date of its lines.

Currently the field is not working as it raises and AttributeError because
there is no origin on the invoice model (but it is on the sale model). This
commit also fixes this error.